Create :mail:protocols:webdav Gradle module
