+++ mail/common/src/main/java/com/fsck/k9/mail/transport/smtp/SmtpTransportUriCreator.java
@@ -1,0 +1,74 @@
+package com.fsck.k9.mail.transport.smtp;
+
+
+import java.io.UnsupportedEncodingException;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.net.URLEncoder;
+
+import com.fsck.k9.mail.AuthType;
+import com.fsck.k9.mail.ServerSettings;
+
+
+public class SmtpTransportUriCreator {
+
+    /**
+     * Creates a SmtpTransport URI with the supplied settings.
+     *
+     * @param server
+     *         The {@link ServerSettings} object that holds the server settings.
+     *
+     * @return A SmtpTransport URI that holds the same information as the {@code server} parameter.
+     *
+     * @see com.fsck.k9.mail.store.StoreConfig#getTransportUri()
+     */
+    public static String createSmtpUri(ServerSettings server) {
+        String userEnc = (server.username != null) ?
+                encodeUtf8(server.username) : "";
+        String passwordEnc = (server.password != null) ?
+                encodeUtf8(server.password) : "";
+        String clientCertificateAliasEnc = (server.clientCertificateAlias != null) ?
+                encodeUtf8(server.clientCertificateAlias) : "";
+
+        String scheme;
+        switch (server.connectionSecurity) {
+            case SSL_TLS_REQUIRED:
+                scheme = "smtp+ssl+";
+                break;
+            case STARTTLS_REQUIRED:
+                scheme = "smtp+tls+";
+                break;
+            default:
+            case NONE:
+                scheme = "smtp";
+                break;
+        }
+
+        String userInfo;
+        AuthType authType = server.authenticationType;
+        // NOTE: authType is append at last item, in contrast to ImapStore and Pop3Store!
+        if (authType != null) {
+            if (AuthType.EXTERNAL == authType) {
+                userInfo = userEnc + ":" + clientCertificateAliasEnc + ":" + authType.name();
+            } else {
+                userInfo = userEnc + ":" + passwordEnc + ":" + authType.name();
+            }
+        } else {
+            userInfo = userEnc + ":" + passwordEnc;
+        }
+        try {
+            return new URI(scheme, userInfo, server.host, server.port, null, null,
+                    null).toString();
+        } catch (URISyntaxException e) {
+            throw new IllegalArgumentException("Can't create SmtpTransport URI", e);
+        }
+    }
+
+    private static String encodeUtf8(String s) {
+        try {
+            return URLEncoder.encode(s, "UTF-8");
+        } catch (UnsupportedEncodingException e) {
+            throw new RuntimeException("UTF-8 not found");
+        }
+    }
+}
