+++ mail/common/src/test/java/com/fsck/k9/mail/transport/smtp/SmtpTransportTest.java
@@ -13,7 +13,6 @@ import com.fsck.k9.mail.Message;
 import com.fsck.k9.mail.MessagingException;
 import com.fsck.k9.mail.ServerSettings;
 import com.fsck.k9.mail.ServerSettings.Type;
-import com.fsck.k9.mail.TransportUris;
 import com.fsck.k9.mail.XOAuth2ChallengeParserTest;
 import com.fsck.k9.mail.filter.Base64;
 import com.fsck.k9.mail.helpers.TestMessageBuilder;
@@ -957,7 +956,7 @@ public class SmtpTransportTest {
                 USERNAME,
                 password,
                 CLIENT_CERTIFICATE_ALIAS);
-        String uri = TransportUris.createTransportUri(serverSettings);
+        String uri = SmtpTransportUriCreator.createSmtpUri(serverSettings);
         StoreConfig storeConfig = setupStoreConfigWithTransportUri(uri);
 
         return new TestSmtpTransport(storeConfig, socketFactory, oAuth2TokenProvider, injectedHostname, injectedIP);
