+++ mail/protocols/webdav/src/main/java/com/fsck/k9/mail/store/webdav/ParsedMessageEnvelope.java
