+++ settings.gradle
@@ -4,5 +4,6 @@ include ':mail:k9mail'
 include ':mail:testing'
 include ':mail:protocols:imap'
 include ':mail:protocols:pop3'
+include ':mail:protocols:webdav'
 include ':plugins:HoloColorPicker'
 include ':plugins:openpgp-api-lib:openpgp-api'
